Rework quoting and spelling
Eberban is phonetic, but borrowed words are not.

In light of this, cen and skan have been removed to make
way for with a more nuanced system.

We now distinguish between denoting words, spelling,
sound, and quotes of words.

Starting from the most fundamental we have:

- ecima: the word for a word
- ecema: the word for a letter/grapheme
- eceuma: the word for a sound/phoneme
- ecama: a gramatically correct Eberban string
- ecoma: a string of data

These can be quoted using:

- ci: quote the next word

- ce: quote a string/list of letters
- ceu: quote a string/list of sounds
- both of these are terminated by cei

- ca: a grammatically correct Eberban quote. Terminated by cai.

- co: a quote of arbitrary length
- quote format: co <delimiter>'<quote content>'<delimiter>
where <delimiter> is a native word form

- coi: a foreign quote of arbitrary length (skipped by speaker)

Spelling and uttering are referenced with:

- ei ce bu: E is spelled as A
- ei ceu bu: E is uttered as A

Enabling us to ask questions such as:

- poi gi be ci umia eicebu ba
- pa gi ce u mi i a cei