♻️🎨📱 Refactor index.html to be mobile-first
An all-new mobile-first responsive index.html featuring semantic HTML
elements and a `<h1>Eberban</h1>`!

The goal with this refactor is to make the index.html page more
accessible, so it's easier to show it to somebody on your phone.

I'm sure it can be made more stylish, but that can be a future PR 😁