📱 Redesign root-generator to be mobile-first
## Redesign

This redesign is based on the following:

- We definitely know the transitivity of the word before we go generate
it.
- We roughly know the length of the word before we go generate it.

Use the ◀️▶️ for navigation and 🔁 to refresh roots.

The transitivity, roots, and carousel position is all cached in local
storage so that you can always come back to where you left off, even if
you refresh the page!

## Implementation

I'm using classes for the UI as I needed functionality that executed for
the first-time only (so constructors made sense to me).

I'm using the store2 library for localStorage as I like the clean API.

Shared styles are now in the reset and common CSS files. And the JS
files for root-generator are split into ui/ and logic/ for clarity.

> [!WARNING]
> With this update you may experience black flickering
> when resizing your chrome/chromium-based browser on their desktop.
>
> I don't know what the root cause of this is, but disabling hardware
> acceleration seems to fix things for me.
>
> https://www.reddit.com/r/chrome/comments/b6x8jw/solved_screen_flashingflickering_frequently_on/