🐛 Fix roots not generating on tab manipulation
The 🔁 button was not working after tab duplication or closing the tab
and opening it again.

There was no problem on the dev server only on the built website.

Vite, using esbuild, minifies JS and mangles function names. This was
interfering with the function name lookup in the root_lists.js file.
After minification, the website could not find the random_root function
names.

Because esbuild is much faster than terser, I've decided to stick with
the former and explicitly use the function names when looking them up
again.