👷 Add PR preview to CI
I initially wanted to separate this out into separate workflows. I
learned, however, that workflows and jobs are isolated and don't share
state.

I thought it would be simplest (although perhaps very janky) to have
this all in one job in one workflow and use conditions to control the
CI flow.

I'm not too sure how the concurrency group works, but it was recommended
as a [best practice](https://github.com/rossjrw/pr-preview-action?tab=readme-ov-file#set-a-concurrency-group).

I also started using a GitHub action for deploying to GitHub pages as I
moved from force-pushing to rebasing and the latter was getting tricky
for me.

I did this because PR previews work by adding a folder on gh-pages, so
force-pushing would overwrite this and we'd lose our previews all the
time.