♻️ Render dictionary with JSX and markup functions
This replaces the current rendering functionality with
[the kit method][markup-PR].

This is mostly a refactor, but there are a couple differences:

- Backticks (e.g., `on <namespace> pi <pred> ...`) are rendered.
- `\[` and `\]` are removed from the `en.yaml`. They aren't needed
  because the RegExp for rendering places is specific enough.

It uses Preact to split up the rendering into JSX components, which makes
it more scaleable and, in my view, more readable.

This PR will make it easier to redesign the dictionary to be
mobile-first.

[markup-PR]: https://github.com/eberban/eberban/pull/97

---

Preact was chosen as I needed to convert a list of dictionary entries to
HTML. I tried templating plugins for rollupJS like post-html and
lit-html but I couldn't get them to work. Preact does the job and is much
smaller than React (I don't envision this repo getting large).