✅ Pass lexing a single eberban particle
This does not affect the website as the lexer remains unintegrated. This
PR [continues the test-driven implementation][tdd] of an Eberban lexer.

[tdd]:  https://github.com/eberban/eberban/pull/100

### New terms

> // A word is lexed into a lexeme if it matches one of its combos.
> — the code

A `combo` is made up of `combo_item`s which are either shyllables or
sonorants.

So, in a way, the lexer parses a combo into a lexeme. I found this
approach reduces regex usage and makes it easier to reason about where
lexing errors would pop up.