✨-unintegrated Add new dictionary rendering code (#97)
* ➕ Install react-string-replace

* ✨-unintegrated Add new dictionary rendering code

This change is unintegrated and thus has no effect on the website.

It's my view that the current implementation to render dictionary entries
unreadable:

- It uses raw RegEx which is [too terse to be readable][link]
- It uses HTML in strings which loses HTML syntax highlighting

This change (let's call it "the kit method") aims to be scaleable as well
as readable.

Kits (a kit is a regex pattern and corresponding replacer function) can
be easily added, nested, and removed from the list that the markup
function uses.

[link]: https://softwareengineering.stackexchange.com/a/298575

* 🧑‍💻 Add regex consts to improve readability

* ♻️ Use Regex set for eberban place

* ♻️ Use capture groups to replace string.slice()

* ♻️ regex.js -> regex.ts