➕ Install preact-parser
htmr was getting dinged in the typescript compilation of the node
dependencies. Since I didn't want to install react itself, I decided to
move to another library. And preact-parser works directly with preact!

This package is only tested on node version 16. So a warning is produced
when installing it. I personally think that, because this package is
tightly scoped (just parsing), that this isn't worth worrying about.