♻️ Rewrite react-string-replace and refactor it (#113)
* ➕ Install htmr and preact-render-to-string

We realised that react-string-replace had a limiting syntax. We wanted to
be able to access the captured strings, so we're implementing our own
version of that function.

* ➕ Install @preact/preset-vite

* ➕ Install @types/react

* ➖ Uninstall react-string-replace

* ♻️ Rename regex function ignore to not_in_set

* ♻️ Add new regex consts for characters

* ♻️ Implement react-string-replace ourselves

* 🚚 markup_text.jsx -> markup_text.tsx

* 🚚 Move markup files to markup folder

* ♻️ Create markup/index.tsx

* 🧑‍💻 Make make_replacements consistent with regex APIs

Motivation:

- whole_string was unused
- filtering capture groups is inconsistent with regex APIs. By no longer
doing this, we're separating concerns.

* ➕ Install preact-parser

htmr was getting dinged in the typescript compilation of the node
dependencies. Since I didn't want to install react itself, I decided to
move to another library. And preact-parser works directly with preact!

This package is only tested on node version 16. So a warning is produced
when installing it. I personally think that, because this package is
tightly scoped (just parsing), that this isn't worth worrying about.

* ➖ Uninstall htmr

* ➖ Uninstall @types/react

* 💡 Shorten comment in replace_string_with_jsx

* 🎨 tidy up the code

* 🩹 Fix newlines being swallowed by preact-parser