👷 Create script+CI to check for unique dict IDs
Contributors to the repository will now be able to locally check
dictionary IDs are unique and any pushes and PRs will fail if all
dictionary IDs (in dictionary/en.yaml) are not unique.

With this change we'll be able to rely on each dictionary ID being
unique.

I tried implementing this with mikefarah's yq, however it does not
retain line numbers upon parsing. The "line" command of yq points to a
node position, not a file line number.