♻️ Render dictionary with JSX and markup functions (#98)
* 🩹 Make dictionary's JS compatible with Preact

In the dev server, `Preact inserts this line of code into the referenced
`../legacy/bootstrap-legacy.js`:

```js
import "{...}/eberban/web/node_modules/.vite/deps/preact_debug.js?v=f6e6bdb6";
```

Which throws this error in the console:

```
Uncaught SyntaxError: Cannot use import statement outside a module
(at bootstrap-legacy.js:1:1)
```

This doesn't seem to have any effect on the website, but since I use the
console all the time for debugging, I made this change to remove it.

* ♻️ Render dictionary with JSX and markup functions

This replaces the current rendering functionality with
[the kit method][markup-PR].

This is mostly a refactor, but there are a couple differences:

- Backticks (e.g., `on <namespace> pi <pred> ...`) are rendered.
- `\[` and `\]` are removed from the `en.yaml`. They aren't needed
  because the RegExp for rendering places is specific enough.

It uses Preact to split up the rendering into JSX components, which makes
it more scaleable and, in my view, more readable.

This PR will make it easier to redesign the dictionary to be
mobile-first.

[markup-PR]: https://github.com/eberban/eberban/pull/97

---

Preact was chosen as I needed to convert a list of dictionary entries to
HTML. I tried templating plugins for rollupJS like post-html and
lit-html but I couldn't get them to work. Preact does the job and is much
smaller than React (I don't envision this repo getting large).

* ⚡️ Cache rendered dictionary entries

This nearly-fully restores the dictionary speed had before we started
using components. 

Components are kept as strings as long as possible to keep up the speed.

* 👔 Re-distinguish between trans equiv and sharing

This is a restoration change, no impact on clients.

This was lost when I ported dictionary.js to
dictionary.jsx and so when I rebased on the master
branch I did not get these changes.