🧪🩹👷 Add lexer tests and setup CI (#100)
* ➕ Install vitest

* 🧪 Add failing lexer tests for spaces, particles, & roots

* 🧪 Add specific non-lexing tests for `-`, `[`, `]`

* 🧪 Adjust spec to lex a single eberban space

This makes more sense than treating only space
as non-lexing as space will be lexed between words.

* 📝 Rename vhowel -> mahul & vunit -> shyllable

As discussed on Discord. A mahul is actually one or more consecutive
vowels, optionally including "h". Using this term, it's conciser to
write tests!

* 🩹 Skip tests until implementation

Skip tests so that CI can be added and won't fail
the PR.

The priority is first getting feedback on the tests
and setting up the CI.

* 👷 Add CI for running website tests

* 💚 Mark lex parameter as unused

* 👷 Ensure CI only runs when editing lexer scripts